+++ src/main/java/com/wf/demo/dao/StudentDao.java
@@ -17,7 +17,7 @@ public interface StudentDao {
 
     int deleteStudent(String id);
 
-    List<Student> queryByClassId(Long classId);
+    List<Student> queryByClassId(int classId);
 
-    int countByClassId(Long classId);
+    int countByClassId(int classId);
 }
