+++ src/main/java/com/wf/demo/entity/ClassInfo.java
@@ -1,24 +1,24 @@
 package com.wf.demo.entity;
 
 public class ClassInfo {
-    private Long id;
+    private int id;
     private String grade;
     private int classNumber;
 
     public ClassInfo() {
     }
 
-    public ClassInfo(Long id, String grade, int classNumber) {
+    public ClassInfo(int id, String grade, int classNumber) {
         this.id = id;
         this.grade = grade;
         this.classNumber = classNumber;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
