+++ src/main/java/com/wf/demo/entity/Student.java
@@ -4,23 +4,23 @@ public class Student {
     private String id;
     private String name;
     private String gender;
-    private Long classId;
+    private int classId;
 
     public Student() {
     }
 
-    public Student(String id, String name, String gender, Long classId) {
+    public Student(String id, String name, String gender, int classId) {
         this.id = id;
         this.name = name;
         this.gender = gender;
         this.classId = classId;
     }
 
-    public Long getClassId() {
+    public int getClassId() {
         return classId;
     }
 
-    public void setClassId(Long classId) {
+    public void setClassId(int classId) {
         this.classId = classId;
     }
