+++ src/main/java/com/wf/demo/service/CourseService.java
@@ -1,14 +1,13 @@
 package com.wf.demo.service;
 
 import com.wf.demo.entity.Course;
-import com.wf.demo.entity.CourseList;
 
 import java.util.List;
 
 public interface CourseService {
     List<Course> queryAllCourse();
 
-    Course queryById(Long id);
+    Course queryById(int id);
 
     Course queryByName(String name);
 
@@ -16,6 +15,6 @@ public interface CourseService {
 
     int updateCourse(Course course);
 
-    int deleteByCourse(Long id);
+    int deleteByCourse(int id);
 
 }
